feat: adding possibility to execute multi page api call
