+++ src/cartridges/bm_adyen/cartridge/controllers/AdyenSettings.js
@@ -7,6 +7,7 @@ const AdyenHelper = require('*/cartridge/adyen/utils/adyenHelper');
 const constants = require('*/cartridge/adyen/config/constants');
 const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 const bmHelper = require('*/cartridge/utils/helper');
+const managementApi = require('*/cartridge/scripts/managementApi');
 
 server.get('Start', (_req, res, next) => {
   if (!csrfProtection.validateRequest()) {
@@ -87,25 +88,9 @@ server.post('TestConnection', server.middleware.https, (req, res, next) => {
 
 server.get('GetStores', server.middleware.https, (req, res, next) => {
   try {
-    const service = bmHelper.initializeAdyenService(
-      constants.SERVICE.GETSTORES,
-      'GET',
-    );
-    const merchantAccount = AdyenConfigs.getAdyenMerchantAccount();
-    const callResult = service.call(JSON.stringify({ merchantAccount }));
-
-    if (!callResult.isOk()) {
-      throw new Error('/getStores call failed');
-    }
-    const resultObject = callResult.object;
-    const response = JSON.parse(resultObject.getText());
-    const mappedData = response.data.map((store) => ({
-      id: store.id,
-      reference: store.reference,
-    }));
-
-    bmHelper.saveMetadataField('Adyen_StoreId', mappedData);
-    res.json({ success: true, stores: mappedData });
+    const stores = managementApi.fetchAllStores();
+    bmHelper.saveMetadataField('Adyen_StoreId', stores);
+    res.json({ success: true, stores });
   } catch (error) {
     AdyenLogs.error_log('Error while fetching stores:', error);
     res.json({ success: false });
